(feature/neeraj-user-activities): move to seprate folder
