--- src/pages/private/dashboard/index.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Dashboard() {
-  return (
-    <h1>Dashboard</h1>
-  )
-}
-
-export default Dashboard
