+++ src/router/routes/privateRoutes.js
@@ -1,6 +1,6 @@
 // Export all the private routes
 
-import Dashboard from "pages/private/dashboard";
+import Dashboard from "pages/private/dashboard/Dashboard";
 
 
 export const PrivateRoutes = [
